feat(ui): 同步按钮统一触发，连接与访问配置收进设置
无远端时打开设置同步分组，已配置则选择同步方向。设置侧栏分为通用、同步、访问。文档补齐宽屏左中右与窄屏底栏+sheet（含筛选与批量）。
