--- app/src/features/connections/ConnectionsModal.css
@@ -1,95 +1,0 @@
-.connections-modal-overlay {
-  position: fixed;
-  inset: 0;
-  z-index: 60;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 1rem;
-  background: rgba(0, 0, 0, 0.4);
-}
-
-.connections-modal-overlay--full {
-  padding: 0;
-  align-items: stretch;
-}
-
-.connections-modal {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: 48rem;
-  max-height: min(90vh, 45rem);
-  overflow: hidden;
-  border-radius: 0.75rem;
-  background: #fff;
-  box-shadow: 0 20px 40px rgba(15, 23, 42, 0.18);
-}
-
-.connections-modal--full {
-  max-width: none;
-  max-height: none;
-  height: 100%;
-  border-radius: 0;
-}
-
-.connections-modal-header,
-.connections-modal-footer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.75rem;
-  padding: 1rem 1.25rem;
-  border-bottom: 1px solid #e5e7eb;
-}
-
-.connections-modal-footer {
-  justify-content: flex-end;
-  border-bottom: none;
-  border-top: 1px solid #e5e7eb;
-}
-
-.connections-modal-heading {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.connections-modal-heading h2 {
-  margin: 0;
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: #111827;
-}
-
-.connections-modal-close {
-  display: inline-flex;
-  min-width: 2.75rem;
-  min-height: 2.75rem;
-  align-items: center;
-  justify-content: center;
-  border: none;
-  border-radius: 0.5rem;
-  background: transparent;
-  color: #6b7280;
-  cursor: pointer;
-}
-
-.connections-modal-body {
-  flex: 1;
-  min-height: 0;
-  overflow: auto;
-  padding: 1.25rem 1.5rem;
-}
-
-.connections-modal-done {
-  min-height: 2.75rem;
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 0.5rem;
-  background: #2563eb;
-  color: #fff;
-  font-size: 0.875rem;
-  font-weight: 500;
-  cursor: pointer;
-}
